+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### State Machine Breaking
 
 * [#2171](https://github.com/crypto-org-chain/cronos/pull/2171) fix(cronos): reject `MsgTurnBridge` instead of reporting success for a no-op.
+* [#2191](https://github.com/crypto-org-chain/cronos/pull/2191) chore(app): remove interchain-accounts (ICA) wiring.
 
 ### Bug fixes